pyproject.toml
[project]
name = "curio"
version = "0.1.0"
description = "Semantic labeling of iMsgX artifacts in preparation for a knowledgebase."
readme = "README.md"
requires-python = ">=3.12"
authors = [
    { name = "Zeph Grunschlag", email = "zephaniah@zibby.com" },
]
keywords = ["cli", "curation", "imsgx", "knowledgebase", "translation"]
classifiers = [
    "Environment :: Console",
    "Operating System :: MacOS :: MacOS X",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: Implementation :: CPython",
]
dependencies = [
    "google-cloud-documentai>=3.14.0",
    "jsonschema>=4.26.0",
    "keyring>=25.7.0",
    "openai>=2.32.0",
    "typer>=0.24.1",
]

[project.urls]
Homepage = "https://github.com/tzaffi/curio"
Repository = "https://github.com/tzaffi/curio"
Issues = "https://github.com/tzaffi/curio/issues"

[dependency-groups]
dev = [
    "pytest>=9.0.3",
    "pytest-cov>=7.1.0",
    "ruff>=0.15.10",
    "ty>=0.0.29",
]

[project.scripts]
curio = "curio.cli:app"

[tool.pytest.ini_options]
addopts = [
    "--strict-config",
    "--strict-markers",
    "--cov=curio",
    "--cov-report=term-missing",
]
markers = [
    "live_codex_cli: skipped by default; set CURIO_LIVE_CODEX_CLI_TESTS=1 to run Codex CLI live smoke tests",
    "live_codex_cli_textify: skipped by default; set CURIO_LIVE_CODEX_CLI_TEXTIFY_TESTS=1 to run Codex CLI textify live smoke tests",
]
testpaths = ["tests"]

[tool.coverage.run]
source = ["curio"]

[tool.coverage.report]
fail_under = 100
show_missing = true

[tool.ruff]
src = ["src", "tests"]
target-version = "py312"

[tool.ruff.lint]
ignore = ["B008", "B019"]
select = ["E4", "E7", "E9", "F", "I", "B", "UP"]

[tool.ty.environment]
python-version = "3.12"

[tool.ty.src]
include = ["src/curio"]

[tool.ty.terminal]
error-on-warning = true

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/curio"]
